Turn the FlaggedRevs check into a filter and use it on relevant content
Make the check a filter and use it to check relevant content:
`/html/{title}`, `/mobile-sections/{title}` and `/pdf/{title}`.

This version is applied only on content that does not specify a revision
in order not to break the editing workflow.
