+++ v1/mobileapps.yaml
@@ -8,6 +8,7 @@ paths:
       - path: ./lib/revision_table_access_check_filter.js
         options:
           redirect_cache_control: '{{options.purged_cache_control}}'
+      - path: lib/flagged_revs_filter.js
     get:
       tags:
         - Mobile
@@ -107,6 +108,7 @@ paths:
       - path: ./lib/revision_table_access_check_filter.js
         options:
           redirect_cache_control: '{{options.purged_cache_control}}'
+      - path: lib/flagged_revs_filter.js
     get:
       tags:
         - Mobile
@@ -191,6 +193,7 @@ paths:
       - path: ./lib/revision_table_access_check_filter.js
         options:
           redirect_cache_control: '{{options.purged_cache_control}}'
+      - path: lib/flagged_revs_filter.js
     get:
       tags:
         - Mobile
