+++ v1/pdf.yaml
@@ -17,6 +17,7 @@ paths:
       - path: ./lib/revision_table_access_check_filter.js
         options:
           redirect_cache_control: '{{options.cache_control}}'
+      - path: lib/flagged_revs_filter.js
     get:
       tags:
         - Page content
